requirements.txt
Flask==2.0.2
Werkzeug==2.0.3
flask-mysqldb
mysql-connector-python
